Golang is bitching about using k as a float64

+++ stat/poisson.go
@@ -89,7 +89,7 @@ func Poisson_CDF(λ float64) func(k int64) float64 {
 
 func Poisson_CDF_a(λ float64) func(k int64) float64 { // analytic solution, less precision
 	return func(k int64) float64 {
-		p := math.Exp(math.Log(IΓint(k+1, λ)) - (LnFact(k)))
+		p := math.Exp(math.Log(IΓint(k+1, λ)) - (LnFact(float64(k))))
 		return p
 	}
 }
@@ -102,6 +102,6 @@ func Poisson_CDF_At(λ float64, k int64) float64 {
 func LnPoisson_CDF_a(λ float64) func(k int64) float64 { // analytic solution, less precision
 	return func(k int64) float64 {
 		k1 := (float64)(k + 1)
-		return log(IΓ(k1, λ)) - LnFact(k)
+		return log(IΓ(k1, λ)) - LnFact(float64(k))
 	}
 }
